Moved the stlstuff.py file into ice4 (from ice3), and restored the reference to it in BScatter_analyzeroughness3. So that means BScatter_analyzeroughness3 can write an stl file, provide the flag Iwantstl=True is set.
I also installed numpy-stl, and tried out some of its functions in plotstl.ipynb. But it's not good for rendering objects as volumes -- I managed to get plotstl.ipynb to make a mesh plot, that's about all. So, numpy-stl might be useful down the road but it doesn't do much now.