Major revision of the retrieval code, formerly called BScatter_retrieve, now called GNBF_retrieve. Calibration and segment information is now held in namelists.
The data are now held in a common folder, crystals. Inside crystals are subfolders that identify a given crystal, with a calibration namelist. Deeper subfolders identify different surfaces of the crystal; each has a segment namelist (or possibly several), and an npz file that holds the solution.

BScatter_analyzeroughness was modified to accommodate the slight differences in the npz solution files, so that it could be run, but it's still pretty ugly.